removing redundant steps and addressing comments
