+++ metadata-ws/src/test/java/uk/ac/ebi/ampt2d/metadata/MetadataWSIntegrationTest.java
@@ -24,14 +24,7 @@ import org.junit.runner.RunWith;
 @RunWith(Cucumber.class)
 @CucumberOptions(
         features = {
-                "src/test/resources/features/analysis.feature",
-                "src/test/resources/features/common.feature",
-                "src/test/resources/features/file.feature",
-                "src/test/resources/features/reference-sequence.feature",
-                "src/test/resources/features/sample.feature",
-                "src/test/resources/features/study.feature",
-                "src/test/resources/features/taxonomy.feature",
-                "src/test/resources/features/web-resource.feature"
+                "src/test/resources/features"
         },
         strict = true,
         plugin = {
